Inflatable routes/model/controllers/index view
