+++ app/controllers/users_controller.rb
@@ -1,7 +1,4 @@
 class UsersController < ApplicationController
-
   def method_name
-
   end
-
 end
