+++ app/models/user.rb
@@ -1,10 +1,7 @@
 class User < ApplicationRecord
+  has_many :inflatables, dependent: destroy
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :validatable
-
-
-
-
 end
