+++ app/views/inflatables/index.html.erb
@@ -1,0 +1,1 @@
+<h1>Inflatable index</h1>
