+++ db/schema.rb
@@ -14,6 +14,17 @@ ActiveRecord::Schema[7.0].define(version: 2023_02_27_153510) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
+  create_table "inflatables", force: :cascade do |t|
+    t.string "title"
+    t.text "description"
+    t.string "category"
+    t.integer "price"
+    t.bigint "user_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["user_id"], name: "index_inflatables_on_user_id"
+  end
+
   create_table "users", force: :cascade do |t|
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
@@ -28,4 +39,5 @@ ActiveRecord::Schema[7.0].define(version: 2023_02_27_153510) do
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
 
+  add_foreign_key "inflatables", "users"
 end
